Use atomic int64s in http code
Undo the workaround of using intptr atomics and accepting that progress
metrics would be wrong for >4GB transfers on 32-bit.

We have real int64 atomics now.

+++ cxhttp/httpinit.c
@@ -83,7 +83,7 @@ bool _httpAppendBytes(strhandle out, const uint8* data, size_t len)
 }
 
 _Use_decl_annotations_
-bool _httpProgressDue(uintptr done, uintptr* seen, size_t interval, bool final)
+bool _httpProgressDue(uint64 done, uint64* seen, size_t interval, bool final)
 {
     if (done == *seen)
         return false;   // nothing new to report, including the final call after a full one
